+++ src/app/[locale]/me/dashboard/monthly-expense/components/monthly-expenses-dialog/monthly-expenses-dialog.tsx
@@ -55,7 +55,7 @@ export function MonthlyExpensesDialog() {
 
     const rawData = {
       label: formData.get("label") as string,
-      amount: parseInt(formData.get("amount") as string),
+      amount: parseFloat(formData.get("amount") as string),
       tags: formData.get("tags") as string,
       notes: formData.get("notes") as string,
       installments: parseInt(formData.get("installments") as string),
